wipro/enterprise_ai_system/requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Frontend
streamlit==1.28.1
plotly==5.18.0

# Data Processing
pandas==2.3.3
numpy==1.26.4
scipy==1.12.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.1

# Time Series
statsmodels==0.14.0
pmdarima==2.0.4

# Reinforcement Learning
stable-baselines3==2.2.1

# Database
supabase==2.3.5
python-dotenv==1.0.0

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.9

# Generative AI
openai==1.3.8

# Model Serialization
joblib==1.3.2

# Utilities
requests==2.31.0
pydantic==2.5.0
python-jose==3.3.0

# Logging & Monitoring
python-logging==2.29
